% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{aq_standardize}
\alias{aq_standardize}
\title{Rescale AQ score given specific mean and variance}
\usage{
aq_standardize(x, mu = 16.94, sigma = 5.59)
}
\arguments{
\item{x}{Original AQ score}

\item{mu}{Population mean}

\item{sigma}{Population standard deviation}
}
\value{
\link{numeric} vector with attributes to convert back to original
}
\description{
Use mean and standard deviation from the non-clinical population.
}
\details{
Resulting scale will have a mean of 0.5 with most of the mass between 0 and
1.
}
\examples{
aq_standardize(25)
}
\seealso{
\href{https://molecularautism.biomedcentral.com/articles/10.1186/2040-2392-6-2}{Ruzich, E. et. al. (2014)}
}
