% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{import_sre}
\alias{import_sre}
\title{Import SRE dataset}
\usage{
import_sre(from_csv = FALSE, raw = FALSE)
}
\arguments{
\item{from_csv}{Load from CSV or RData. \code{Logical}.}

\item{raw}{Import data without refactoring or standardizing.}
}
\value{
\link{data.frame}
}
\description{
Import SRE dataset
}
\examples{
import_sre()         # load from compressed RData
import_sre(TRUE)     # reload from package CSV data
import_sre(raw=TRUE) # load without refactoring and scaling.
}
