% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{posterior_aq_action_contrasts}
\alias{posterior_aq_action_contrasts}
\title{Univariate form with renamed variables}
\usage{
posterior_aq_action_contrasts(ppd)
}
\description{
Univariate form with renamed variables
}
