% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_fit.R
\name{sre_models}
\alias{sre_models}
\title{Load SRE models into global workspace}
\usage{
sre_models(refit = FALSE, root_dir = NULL, debug = FALSE)
}
\arguments{
\item{refit}{If \code{TRUE}, then refit package models and save to \code{root_dir}. If
\code{FALSE}, then load models from \code{root_dir}.}

\item{root_dir}{Root directory of subfolders containing saved models.
Defaults to external package data directory.}

\item{debug}{Runs a small number of iterations for debugging purposes.}
}
\value{
\link{NULL}
}
\description{
Load SRE models into global workspace
}
\examples{
sre_models()
}
