% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRE-package.R
\docType{package}
\name{SRE-package}
\alias{SRE-package}
\title{Self-Recognition Experiment (SRE)}
\description{
Descrip
}
\details{
See \code{help(package = "SRE")} for a list of functions.

View the vignette with \code{browseVignettes(package ="SRE")}.
}
