% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRE-package.R
\docType{data}
\name{sre}
\alias{sre}
\title{Self-Recognition Experiment (SRE) dataset}
\format{Dataset is an object of class \link{data.frame}.}
\usage{
sre
}
\value{
\code{sre} returns an object of class \link{data.frame}.
}
\description{
Self-Recognition Experiment (SRE) dataset
}
\section{Variables}{

\itemize{
\item \code{recog_acc}: self-recognition accuracy
}
}

\examples{
# get data
library(SRE)
sre_copy <- sre
}
\seealso{
\code{\link[=import_sre]{import_sre()}}
}
\keyword{datasets}
